requirements.txt
torch>=1.12
torchvision
onnx
onnxruntime
fastapi
uvicorn
pydantic
opencv-python
albumentations
numpy
Pillow
tqdm
pyyaml
rq
redis
